# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Disaster Response and Resource Allocation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List all incidents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, monitoring, critical]
        - name: category
          in: query
          schema:
            type: string
            enum: [flood, earthquake, cyclone, fire, landslide]
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create a new disaster incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    delete:
      operationId: deleteIncident
      tags: [incidents]
      summary: Delete incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List all resources
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: createResource
      tags: [resources]
      summary: Add resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceInput"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    patch:
      operationId: updateResource
      tags: [resources]
      summary: Update resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/allocate:
    post:
      operationId: allocateResources
      tags: [resources]
      summary: Run greedy resource allocation for an incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocationRequest"
      responses:
        "200":
          description: Allocation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationResult"
...
  /rescue-teams:
    get:
      operationId: listRescueTeams
      tags: [rescue-teams]
      summary: List rescue teams
      responses:
        "200":
          description: Team list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RescueTeam"
    post:
      operationId: createRescueTeam
      tags: [rescue-teams]
      summary: Register a rescue team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescueTeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RescueTeam"
  /rescue-teams/{id}:
    patch:
      operationId: updateRescueTeam
      tags: [rescue-teams]
      summary: Update rescue team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescueTeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RescueTeam"

  # ── Routes ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Incident:
      type: object
      required: [id, title, category, severity, status, latitude, longitude, affectedPopulation, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [flood, earthquake, cyclone, fire, landslide]
        severity:
          type: integer
          minimum: 1
          maximum: 10
        status:
          type: string
          enum: [active, resolved, monitoring, critical]
        latitude:
          type: number
        longitude:
          type: number
        affectedPopulation:
          type: integer
        imageUrl:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    IncidentInput:
      type: object
      required: [title, category, severity, latitude, longitude, affectedPopulation]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [flood, earthquake, cyclone, fire, landslide]
        severity:
          type: integer
          minimum: 1
          maximum: 10
        latitude:
          type: number
        longitude:
          type: number
        affectedPopulation:
          type: integer
        imageUrl:
          type: string
        region:
          type: string
        status:
          type: string
          enum: [active, resolved, monitoring, critical]
    IncidentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: integer
          minimum: 1
          maximum: 10
        status:
          type: string
          enum: [active, resolved, monitoring, critical]
        affectedPopulation:
          type: integer
        imageUrl:
          type: string
...
    Resource:
      type: object
      required: [id, name, category, totalQuantity, availableQuantity, unit, location]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [food, water, medical, rescue_team, boat, ambulance]
        totalQuantity:
          type: integer
        availableQuantity:
          type: integer
        unit:
          type: string
        location:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
    ResourceInput:
      type: object
      required: [name, category, totalQuantity, availableQuantity, unit, location]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [food, water, medical, rescue_team, boat, ambulance]
        totalQuantity:
          type: integer
        availableQuantity:
          type: integer
        unit:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    ResourceUpdate:
      type: object
      properties:
        availableQuantity:
          type: integer
        totalQuantity:
          type: integer
        location:
          type: string
    AllocationRequest:
      type: object
      required: [incidentId, severity, affectedPopulation]
      properties:
        incidentId:
          type: integer
        severity:
          type: integer
        affectedPopulation:
          type: integer
    AllocationResult:
      type: object
      required: [incidentId, allocations, efficiencyScore, shortages]
      properties:
        incidentId:
          type: integer
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/ResourceAllocation"
        efficiencyScore:
          type: number
        shortages:
          type: array
          items:
            type: string
    ResourceAllocation:
      type: object
      required: [resourceId, resourceName, category, allocated, unit]
      properties:
        resourceId:
          type: integer
        resourceName:
          type: string
        category:
          type: string
        allocated:
          type: integer
        unit:
          type: string
...
    RescueTeam:
      type: object
      required: [id, name, type, size, status, latitude, longitude]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [medical, firefighting, flood_rescue, search_rescue, logistics]
        size:
          type: integer
        status:
          type: string
          enum: [available, deployed, resting, standby]
        latitude:
          type: number
        longitude:
          type: number
        assignedIncidentId:
          type: ["integer", "null"]
        equipment:
          type: array
          items:
            type: string
    RescueTeamInput:
      type: object
      required: [name, type, size, latitude, longitude]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [medical, firefighting, flood_rescue, search_rescue, logistics]
        size:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        equipment:
          type: array
          items:
            type: string
    RescueTeamUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [available, deployed, resting, standby]
        assignedIncidentId:
          type: ["integer", "null"]
        latitude:
          type: number
        longitude:
          type: number

    # ── Routes ────────────────────────────────────────────────────────────────
...