# lib/api-zod/src/generated/resource.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Api
 * Smart Disaster Response and Resource Allocation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceCategory } from './resourceCategory';
import type { ResourceLatitude } from './resourceLatitude';
import type { ResourceLongitude } from './resourceLongitude';

export interface Resource {
  availableQuantity: number;
  category: ResourceCategory;
  id: number;
  latitude?: ResourceLatitude;
  location: string;
  longitude?: ResourceLongitude;
  name: string;
  totalQuantity: number;
  unit: string;
}