# lib/api-zod/src/generated/resourceAllocation.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Api
 * Smart Disaster Response and Resource Allocation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceAllocation {
  allocated: number;
  category: string;
  resourceId: number;
  resourceName: string;
  unit: string;
}