# lib/api-zod/src/generated/resourceUpdate.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Api
 * Smart Disaster Response and Resource Allocation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceUpdate {
  availableQuantity?: number;
  location?: string;
  totalQuantity?: number;
}